+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2
+
+* Change feePayerAccount to connectedAccount.
+* Fix AuthorizeResult walletUriBase serialization error.
+* Add skipReauthorize option to privileged method calls.
+* Add openUri and isWalletInstalled platform methods.
+
 ## 0.0.1+3
 
 * Require max value for randomValue.